Restructure test suite
